Send data on a TCP connection.
Has:
- Validation of JSON data
- Simple filtering (we should do this based on event)

The concept here works, and we can receive it via telnet as a client. All works as expected
I'm sure there's some issues with the sockets, but for now it seems to work as expected.

The filtering used here is very simple, and i believe that itmight actually be nice to create a program that does the same, but pushes the mesages to MQTT or something like that. Because you'd then be able to filter it based on the client's subscription.