Add location_logger.
Simple implementation of the reading side of the dir_processor.

Had to implement a simple protocol to tell the client how large the message would be.
So we're always sending 2 bytes (as uint16) /before/ a message. We then wait for the message with exactly the right size.

We do that in a loop, and then thigns work as expected.

Still lots to do on this, but the basics are there.