Get protobufs going.
I'm not sure about this, but it looks like go is really picky on where you put your files.
I have a loramsgs struct that i use to deserialize the binary array. I'd want to store that in the same directory as the protobufs i just made, but that's not possible. So this is the first commit to clean some of that data up.